requirements.txt
Django>=3.0,<4.0
psycopg2>=2.8
djangorestframework>=3.6
django-filter>=21.0
requests>=2.0
beautifulsoup4>=4.0
boto3>=1.2
markdown>=3.3.0
django-rest-auth>=0.9.0
django-crispy-forms>=1.11.1
